+++ test/expect/climate_non_ret.mli
@@ -1,0 +1,5 @@
+(** This is a helper that allows printing climate errors in the expect tests. *)
+
+type t = Climate.For_test.Non_ret.t
+
+val print : t -> unit